requirements.txt
connexion
pyyaml
Jinja2
-e git+https://github.com/icclab/docker-py.git@695cd770ec4c98070493fa563eb7b298b2f913e2#egg=docker